backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
stripe==7.4.0
solana==0.30.2
solders==0.18.1
httpx==0.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
boto3==1.29.7
pytest==7.4.3
pytest-asyncio==0.21.1
web3==6.15.1
authlib==1.2.1
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
email-validator==2.1.0
cryptography==41.0.7
pycryptodome==3.19.0
python-gnupg==0.5.1
schedule==1.2.0
pyotp==2.9.0
qrcode[pil]==7.4.2
python3-saml==1.15.0
onelogin==1.0.0

